don't use os.exec on linux for clearing screen and move to util package
